+++ src/SendIt/src/MailQueue.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\SendIt;
@@ -14,22 +7,17 @@ namespace Spiral\SendIt;
 use Spiral\Mailer\MailerInterface;
 use Spiral\Mailer\MessageInterface;
 use Spiral\Queue\Options;
+use Spiral\Queue\QueueInterface;
 use Spiral\SendIt\Config\MailerConfig;
 
 final class MailQueue implements MailerInterface
 {
     public const JOB_NAME = 'sendit.mail';
 
-    /** @var MailerConfig */
-    private $config;
-
-    /** @var \Spiral\Queue\QueueInterface */
-    private $queue;
-
-    public function __construct(MailerConfig $config, $queue)
-    {
-        $this->config = $config;
-        $this->queue = $queue;
+    public function __construct(
+        private readonly MailerConfig $config,
+        private readonly QueueInterface $queue
+    ) {
     }
 
     public function send(MessageInterface ...$message): void